requirements.txt
requests
Pillow
matplotlib
pysmhi
aiohttp
PyYAML
astral
IPy
fastapi
uvicorn
jinja2
python-multipart
paho-mqtt